Add files via upload
Hey, so most things are working now except for the server-side temperature and timestamp, all the buttons (digital and physical) are working correctly. Oh, and there's a "synchronize all clients" function that is working right now, it just needs the UI functions to sync the UI controls with the current fan and pump states.